+++ lib/event_api.dart
@@ -1,6 +1,7 @@
 /// Holds APIRepository and implementations of a requester interface
 library event_api;
 
+export 'src/database.dart';
 export 'src/fake.dart';
 export 'src/model.dart';
 export 'src/repository.dart';