+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Added ApiDatabaseRepository - an implementation of DatabaseRepository that is backed by an APIRequester
+- Added StubAPIRequester - an implementation of APIRequester that will always return the same response regardless of urlSuffix
+
 ## 0.1.0+1
 
 - Fixed exports to export all the libraries